+++ src/main/java/com/epam/tc/page/DiffElementsPage/DiffElementsLogWindow.java
@@ -1,5 +1,6 @@
-package com.epam.tc.hw3.page.DiffElementsPage;
+package com.epam.tc.page.DiffElementsPage;
 
+import io.qameta.allure.Step;
 import org.assertj.core.api.Assertions;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -17,7 +18,7 @@ public class DiffElementsLogWindow extends DiffElementsPage {
     @FindBy(xpath = "//ul[contains(@class,'logs')]//li")
     List<WebElement> logList;
 
-
+    @Step("Check text and position record in Log Window")
     public void checkWebElementInLogWindow(int positionElementInLogWindow, String textElementInLogWindow) {
         Assertions.assertThat(logList.get(positionElementInLogWindow - 1).getText().contains(textElementInLogWindow)).isTrue();
     }
