+++ src/main/java/com/epam/tc/page/IndexPage/IndexPageHeader.java
@@ -1,4 +1,4 @@
-package com.epam.tc.hw4.page.IndexPage;
+package com.epam.tc.page.IndexPage;
 
 import io.qameta.allure.Step;
 import org.assertj.core.api.Assertions;
@@ -41,7 +41,7 @@ public class IndexPageHeader extends IndexPage {
         Assertions.assertThat(menuMetCol.isDisplayed());
     }
 
-    @Step("Open DiffElement Page")
+    @Step("Open DiffElements Page")
     public void openDiffElementPage() {
         menuServ.click();
         servDiffElements.click();
