+++ src/main/java/com/epam/tc/util/AttachmentUtil.java
@@ -1,4 +1,4 @@
-package com.epam.tc.hw4.util;
+package com.epam.tc.util;
 
 import io.qameta.allure.Allure;
 import io.qameta.allure.Attachment;
@@ -8,7 +8,7 @@ import java.io.InputStream;
 
 @UtilityClass
 public class AttachmentUtil {
-
+    @Attachment
     public void attachFromInputStream(final String name, InputStream inputStream) {
         Allure.addAttachment(name, inputStream);
     }
