+++ src/test/java/com/epam/tc/hw4/failedTests/FailedTest.java
@@ -1,10 +1,12 @@
 package com.epam.tc.hw4.failedTests;
 
-import com.epam.tc.hw4.page.BasePage;
-import com.epam.tc.hw4.page.IndexPage.IndexPage;
-import com.epam.tc.hw4.page.IndexPage.IndexPageHeader;
-import com.epam.tc.hw4.page.IndexPage.IndexPageLeftMenu;
+import com.epam.tc.page.BasePage;
+import com.epam.tc.page.IndexPage.IndexPage;
+import com.epam.tc.page.IndexPage.IndexPageHeader;
+import com.epam.tc.page.IndexPage.IndexPageLeftMenu;
 import org.testng.annotations.Test;
+
+import static com.epam.tc.util.PropertiesUtil.getValue;
 
 public class FailedTest extends BasePage {
 
@@ -13,7 +15,7 @@ public class FailedTest extends BasePage {
 
         IndexPage indexPage = new IndexPage(driver);
         //1. Open test site by URL
-        indexPage.openPage(propUtil.getValue("url.index"));
+        indexPage.openPage(getValue("url.index"));
         //2. Assert Browser title
         indexPage.checkTitle("Home Pag");
         //3. Perform login
