--- src/test/java/com/epam/tc/hw4/util/PropertiesUtil.java
@@ -1,35 +1,0 @@
-package com.epam.tc.hw4.util;
-
-import io.qameta.allure.Step;
-
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.Properties;
-
-public class PropertiesUtil {
-
-    Properties prop = new Properties();
-    FileInputStream fis = null;
-
-    @Step
-    public String getValue(String key) {
-        try {
-            fis = new FileInputStream("./src/test/resources/application.property");
-            prop.load(fis);
-            return prop.getProperty(key);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            try {
-                fis.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-        return null;
-    }
-
-}
