+++ src/test/resources/testng/seleniumHW3Suite.xml
@@ -2,8 +2,8 @@
 <suite name="SeleniumHW3Suite">
     <test name="SeleniumHW3Test">
         <classes>
-            <class name="com.epam.tc.hw3.ex1.Exercise1"/>
-            <class name="com.epam.tc.hw3.ex2.Exercise2"/>
+            <class name="com.epam.tc.hw3.ex1.UiIndexPageTests"/>
+            <class name="com.epam.tc.hw3.ex2.UiDiffElementsPageTests"/>
         </classes>
     </test>
 </suite>
