+++ src/test/resources/testng/seleniumHW4AndFailedTestSuite.xml
@@ -5,8 +5,8 @@
     </listeners>
     <test name="SeleniumHW4AndFailedTest">
         <classes>
-            <class name="com.epam.tc.hw4.ex1.Exercise1"/>
-            <class name="com.epam.tc.hw4.ex2.Exercise2"/>
+            <class name="com.epam.tc.hw4.ex1.UiIndexPageTests"/>
+            <class name="com.epam.tc.hw4.ex2.UiDiffElementsPageTests"/>
             <class name="com.epam.tc.hw4.failedTests.FailedTest"/>
         </classes>
     </test>
